Move tracer options to lib/tracer
Its not clear why they were in  trace-object.js, they aren't related to
the rest of that file.
