+++ bin/sl-run.js
@@ -56,8 +56,7 @@ if ((config.clustered && config.isMaster) || config.detach) {
 
 if (config.enableTracing) {
   var tracer = require('../lib/tracer');
-  var traceObject = require('../lib/trace-object');
-  tracer(traceObject.tracerOptions());
+  tracer(tracer.tracerOptions());
 }
 
 config.sendMetrics();
