+++ lib/trace-object.js
@@ -1,16 +1,6 @@
-var agent = require('./agent');
 var tracer = require('./tracer');
 var cluster = require('cluster');
 var debug = require('./debug')('trace-object');
-
-function tracerOptions() {
-  return {
-    archiveInterval: 20000,
-    accountKey: agent().config.appName,
-    useHttp: false
-  };
-}
-exports.tracerOptions = tracerOptions;
 
 exports.sendTraceObject = function sendTraceObject(parentCtl) {
   if (cluster.isWorker) {
