+++ lib/tracer.js
@@ -1,4 +1,7 @@
-module.exports = tracer;
+var agent = require('./agent');
+
+exports = module.exports = tracer;
+exports.tracerOptions = tracerOptions;
 
 var TRACER = null;
 
@@ -7,4 +10,12 @@ function tracer(options) {
     TRACER = require('strong-trace')(options);
   }
   return TRACER;
+}
+
+function tracerOptions() {
+  return {
+    archiveInterval: 20000,
+    accountKey: agent().config.appName,
+    useHttp: false
+  };
 }
